requirements.txt
Django==2.1.7
django-configurations==2.1
django-cors-headers==2.4.0
django-extensions==2.1.6
django-filter==2.1.0
djangorestframework==3.9.1
Markdown==3.0.1
drf-yasg==1.13.0
mysqlclient==1.4.2.post1
pytz==2018.9
six==1.12.0
swagger-spec-validator==2.4.3